EID-125: Fix JS test failures & infinite loop issue
We had two weird problems with our JS tests.  When you ran the tests on
their own they would pass, but when you ran them with the other code
they'd fail, and the jasmine runner would get stuck in an infinite loop,
repeatedly rerunning all the tests.

There seem to have been two problems here:

1) Our tests weren't careful enough when attaching submit handlers that
   the submit event was prevented from bubbling up to the document and
   triggering a reload of the jasmine tests.
2) There were some sprockets comments in some othe file that *for some
   unknown reason* were causing the tests to behave differently.

We've addressed the first issue by:

* Switching to attaching `submit` handlers with `.on('submit', ...)`
  instead of `.submit(...)` for clarity.
* Being more precise about the form that we're triggering the submit
  event on (previously we were accidentally submitting both of the forms
  on in the `$dom` sandbox).
* Returning `false` in the submit handlers to prevent the submit event
  from bubbling up to the document and reloading the page.
* Only attaching the handlers that check the submit happened correctly
  to the `$dom` sandbox, instead of the `document`. This ensures they're
  properly cleaned up in the `afterEach` when we call `$dom.remove()`.

We've addressed the second issue by removing the offending sprockets
comments. These comments don't seem to be needed to make the tests pass
now (we've checked running the tests in isolation and alongside the
other tests). The theory is that these are a hangover from the old
rails-jasmine framework we used to use in the olden days. In the new
framework the dependencies they're asking for have already been loaded,
so the comments break things by trying to load them twice (we haven't
been able to confirm this though).

All-in-all this investigation has been very frustrating and difficult to
conduct. This was exacerbated by the inability to debug the tests, which
we used to be able to do by loading the tests in a browser and using the
dev tools to debug. Since we switched to the new jasmine runner we
haven't been able to do this.

Authors: @richardtowers